+++ airbyte-integrations/base-python/.gitignore
@@ -1,1 +1,1 @@
-airbyte_protocol/models
+airbyte_protocol/models/yaml
