+++ airbyte-integrations/base-python/airbyte_protocol/logger.py
@@ -1,0 +1,40 @@
+from .models import AirbyteLogMessage, AirbyteMessage
+
+
+class AirbyteLogger:
+    def __init__(self):
+        self.valid_log_types = ["FATAL", "ERROR", "WARN", "INFO", "DEBUG", "TRACE"]
+
+    def log_by_prefix(self, message, default_level):
+        split_line = message.split()
+        first_word = next(iter(split_line), None)
+        if first_word in self.valid_log_types:
+            log_level = first_word
+            rendered_message = " ".join(split_line[1:])
+        else:
+            log_level = default_level
+            rendered_message = message
+        self.log(log_level, rendered_message)
+
+    def log(self, level, message):
+        log_record = AirbyteLogMessage(level=level, message=message)
+        log_message = AirbyteMessage(type="LOG", log=log_record)
+        print(log_message.serialize())
+
+    def fatal(self, message):
+        self.log("FATAL", message)
+
+    def error(self, message):
+        self.log("ERROR", message)
+
+    def warn(self, message):
+        self.log("WARN", message)
+
+    def info(self, message):
+        self.log("INFO", message)
+
+    def debug(self, message):
+        self.log("DEBUG", message)
+
+    def trace(self, message):
+        self.log("TRACE", message)
