+++ airbyte-integrations/base-python/main.py
@@ -1,0 +1,6 @@
+import sys
+
+from airbyte_protocol.entrypoint import main
+
+if __name__ == "__main__":
+    main(sys.argv[1:])
