+++ airbyte-integrations/singer/exchangeratesapi/source/source_exchangeratesapi_singer/source.py
@@ -3,6 +3,7 @@ from typing import Generator
 
 from airbyte_protocol import AirbyteCatalog
 from airbyte_protocol import AirbyteCheckResponse
+from airbyte_protocol.base import AirbyteEntrypoint
 from airbyte_protocol import AirbyteMessage
 from airbyte_protocol import AirbyteSpec
 from airbyte_protocol import Source
@@ -22,10 +23,16 @@ class SourceExchangeRatesApiSinger(Source):
         return AirbyteCheckResponse(code == 200, {})
 
     def discover(self, logger, config_container) -> AirbyteCatalog:
-        catalogs = SingerHelper.get_catalogs(logger, "tap-exchangeratesapi | grep '\"type\": \"SCHEMA\"' | head -1 | jq -c '{\"streams\":[{\"stream\": .stream, \"schema\": .schema}]}'")
+        catalogs = SingerHelper.get_catalogs(logger,
+                                             "tap-exchangeratesapi | grep '\"type\": \"SCHEMA\"' | head -1 | jq -c '{\"streams\":[{\"stream\": .stream, \"schema\": .schema}]}'")
         return catalogs.airbyte_catalog
 
     def read(self, logger, config_container, catalog_path, state=None) -> Generator[AirbyteMessage, None, None]:
         config_option = f"--config {config_container.rendered_config_path}"
         state_option = f"--state {state}" if state else ""
         return SingerHelper.read(logger, f"tap-exchangeratesapi {config_option} {state_option}")
+
+
+if __name__ == "__main__":
+    source = SourceExchangeRatesApiSinger()
+    AirbyteEntrypoint(source).start()
